(0) Practice Calling Object Methods
CALLING METHODS

(1) Create new Android app

Once your project gets created, go ahead and run the app on your device.

The default behavior is for Hello World to show.

Now we'll modify the onCreate method in the main acitivity class.

Instead of calling setContentView on this activity main layout:
setContentView(R.layout.activity_main);

...we're gonna create our own TextView and set that here.

From android developers site:
http://developer.android.com/training/basics/firstapp/starting-activity.html

CTRL F "TextView textView"

It tells you that in the onCreate(), we can create a TextView:

TextView textView = new TextView(this);

We're creating a new TextView and we're stroring it in a variable called textView. This constructor requires a context as an input parameter. And passing (this) as an input parameter, because (this) refers to the current activity.

And an activity is also a context, so this will correctly create a TextView object for us.

Then we can call methods on this TextView object:

textView.setTextSize(40);
textView.setText(message);

Remember that "textView" is a variable because t is small. TextView is an object.

When you scroll down it says to call setContentView and pass in this textView as an input parameter.

We're just focused on this part where we're creating a textView and then we're setting that as the contentView of the activity layout.

Back in the Android Studio I'm gonna modify the onCreate method, like this:

First I'm gonna create a new TextObject using the constructor:

TextView textView = new TextView();

The TextView() constructor requires a context, so I'm going to pass in this to refer to this activity.

TextView textView = new TextView(this);

Then on the next line of code, I'm gonna call a method on the TextView object.

As soon as I hit dot, it's going to showe me all the possible methods that I could call on this textView object. If I start to type then it will filter down these auto complete options.

I'm going to choose setText so I can hit enter to use this auto complete. Then I can put in a string as an input argument.

textView.setText("Wow!");

And the last step is to change the setContentView line, so instead of using this XML
layout file I'm going to just paste in textView.

setContentView(textView);

I'm going to save this and then run this on the device:

@Override
protected void onCreate(Bundle savedInstanceState) {
    super.onCreate(savedInstanceState);

    TextView textView = new TextView(this);
    textView.setText("Wow!");

    setContentView(textView);
}

Actual code:

package com.example.android.myapplication5;

import android.os.Bundle;
import android.support.v7.app.AppCompatActivity;
import android.widget.TextView;

public class MainActivity extends AppCompatActivity {

    @Override
    protected void onCreate(Bundle savedInstanceState) {
        super.onCreate(savedInstanceState);
        TextView textView = new TextView(this);
        textView.setText("Wow!");

        setContentView(textView);
    }

}

What I expect now is that the screen will show a new TextView that says Wow on it instead of Hello World.

If I wanted to change text size, I'd do this:
        TextView textView = new TextView(this);
        textView.setText("Wow!");
        textView.setTextSize(56);

        setContentView(textView);

Then I can change textColor as well: